New faircoin app: initial commit.
