+++ faircoin/daemon/daemon.conf.sample
