+++ faircoin/daemon/daemon.py.sample
