+++ faircoin/daemon/daemon_service.sample
