+++ faircoin/management/commands/reassign-wallet-addresses.py
