+++ faircoin/templates/faircoin/faircoin_history.html
