+++ faircoin/utils.py
@@ -1,13 +1,14 @@
 import requests, json, logging
 from random import randint
+from logging.handlers import TimedRotatingFileHandler
 
 from django.conf import settings
 
 def init_logger():
-    logger = logging.getLogger("faircoins")
+    logger = logging.getLogger("faircoin")
     logger.setLevel(logging.WARNING)
-    fhpath = "/".join([settings.PROJECT_ROOT, "faircoins.log",])
-    fh = logging.handlers.TimedRotatingFileHandler(fhpath, when="d", interval=1, backupCount=7)
+    fhpath = "/".join([settings.PROJECT_ROOT, "faircoin/faircoin.log",])
+    fh = TimedRotatingFileHandler(fhpath, when="d", interval=1, backupCount=7)
     fh.setLevel(logging.DEBUG)
     formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
     fh.setFormatter(formatter)
