+++ work/views.py
@@ -41,8 +41,6 @@ if "pinax.notifications" in settings.INSTALLED_APPS:
 else:
     notification = None
 
-FAIRCOIN_DIVISOR = Decimal("1000000.00")
-
 def get_site_name(request=None):
     if request:
         domain = request.get_host()
@@ -306,299 +304,6 @@ def register_skills(request):
         "skills": skills,
     })
 '''
-
-
-#    F A I R C O I N
-
-@login_required
-def manage_faircoin_account(request, resource_id):
-    if not settings.USE_FAIRCOIN:
-        raise Http404
-    resource = get_object_or_404(EconomicResource, id=resource_id)
-    user_agent = get_agent(request)
-    if not user_agent or not (resource.owner() == user_agent or resource.owner() in user_agent.managed_projects()):
-        raise Http404
-
-    send_coins_form = None
-    is_wallet_address = None
-    limit = 0
-    confirmed_balance = None
-    unconfirmed_balance = None
-    faircoin_account = False
-    payment_due = False
-    share_price = False
-    number_of_shares = False
-    can_pay = False
-
-    wallet = faircoin_utils.is_connected()
-    if wallet:
-        is_wallet_address = faircoin_utils.is_mine(resource.digital_currency_address)
-        if not is_wallet_address:
-            if resource.is_address_requested(): is_wallet_address = True
-        if is_wallet_address:
-            send_coins_form = SendFairCoinsForm(agent=resource.owner())
-            try:
-                balances = faircoin_utils.get_address_balance(resource.digital_currency_address)
-                confirmed_balance = Decimal(balances[0]) / FAIRCOIN_DIVISOR
-                unconfirmed_balance =  Decimal(balances[0] + balances[1]) / FAIRCOIN_DIVISOR
-                unconfirmed_balance += resource.balance_in_tx_state_new()
-                fee = Decimal(faircoin_utils.network_fee()) / FAIRCOIN_DIVISOR
-                limit = min(confirmed_balance, unconfirmed_balance) - fee
-            except:
-                confirmed_balance = "Not accessible now"
-                unconfirmed_balance = "Not accessible now"
-                limit = Decimal("0.0")
-        else:
-            wallet = False
-
-    candidate_membership = resource.owner().candidate_membership()
-    if candidate_membership:
-        faircoin_account = resource.owner().faircoin_resource()
-        if faircoin_account and wallet:
-            share = EconomicResourceType.objects.membership_share()
-            share_price = share.price_per_unit
-            number_of_shares = resource.owner().number_of_shares()
-            share_price = share_price * number_of_shares
-            payment_due = True
-            if resource.owner().owns_resource_of_type(share):
-                payment_due = False
-            if confirmed_balance and confirmed_balance != "Not accessible now":
-                can_pay = confirmed_balance >= share_price
-
-    return render(request, "work/faircoin_account.html", {
-        "resource": resource,
-        "photo_size": (128, 128),
-        "agent": resource.owner(),
-        "wallet": wallet,
-        "send_coins_form": send_coins_form,
-        "is_wallet_address": is_wallet_address,
-        "confirmed_balance": confirmed_balance,
-        "unconfirmed_balance": unconfirmed_balance,
-        "limit": limit,
-        "faircoin_account": faircoin_account,
-        "candidate_membership": candidate_membership,
-        "payment_due": payment_due,
-        "share_price": share_price,
-        "number_of_shares": number_of_shares,
-        "can_pay": can_pay,
-        "help": get_help("profile"),
-    })
-
-def validate_faircoin_address_for_worker(request):
-    if not settings.USE_FAIRCOIN:
-        raise Http404
-    data = request.GET
-    address = data["to_address"].strip()
-    answer = faircoin_utils.is_valid(address)
-    if answer == False:
-        answer = "Invalid FairCoin address"
-    response = simplejson.dumps(answer, ensure_ascii=False)
-    return HttpResponse(response, content_type="text/json-comment-filtered")
-
-@login_required
-def change_faircoin_account(request, resource_id):
-    if not settings.USE_FAIRCOIN:
-        raise Http404
-    if request.method == "POST":
-        resource = get_object_or_404(EconomicResource, pk=resource_id)
-        form = EconomicResourceForm(data=request.POST, instance=resource)
-        if form.is_valid():
-            data = form.cleaned_data
-            resource = form.save(commit=False)
-            resource.changed_by=request.user
-            resource.save()
-            return HttpResponseRedirect('/%s/%s/'
-                % ('work/manage-faircoin-account', resource_id))
-        else:
-            raise ValidationError(form.errors)
-
-@login_required
-def transfer_faircoins(request, resource_id):
-    if not settings.USE_FAIRCOIN:
-        raise Http404
-    if request.method == "POST":
-        resource = get_object_or_404(EconomicResource, id=resource_id)
-        agent = get_agent(request)
-        to_agent = request.POST["to_user"]
-        send_coins_form = SendFairCoinsForm(data=request.POST, agent=resource.owner())
-        if send_coins_form.is_valid():
-            data = send_coins_form.cleaned_data
-            address_end = data["to_address"]
-            quantity = data["quantity"]
-            notes = data["description"]
-            address_origin = resource.digital_currency_address
-            network_fee = faircoin_utils.network_fee()
-            if address_origin and address_end and network_fee:
-                from_agent = resource.owner()
-                to_resources = EconomicResource.objects.filter(digital_currency_address=address_end)
-                to_agent = None
-                if to_resources:
-                    to_resource = to_resources[0] #shd be only one
-                    to_agent = to_resource.owner()
-                et_give = EventType.objects.get(name="Give")
-                if to_agent:
-                    tt = ExchangeService.faircoin_internal_transfer_type()
-                    xt = tt.exchange_type
-                    date = datetime.date.today()
-                    exchange = Exchange(
-                        exchange_type=xt,
-                        use_case=xt.use_case,
-                        name="Transfer Faircoins",
-                        start_date=date,
-                        notes=notes,
-                        #context_agent=from_agent, # don't set it to allow to_agent to see the exchange
-                    )
-                    exchange.save()
-                    transfer = Transfer(
-                        transfer_type=tt,
-                        exchange=exchange,
-                        transfer_date=date,
-                        name="Transfer Faircoins",
-                        notes=notes,
-                        context_agent=from_agent,
-                    )
-                    transfer.save()
-                else:
-                    tt = ExchangeService.faircoin_outgoing_transfer_type()
-                    xt = tt.exchange_type
-                    date = datetime.date.today()
-                    exchange = Exchange(
-                        exchange_type=xt,
-                        use_case=xt.use_case,
-                        name="Send Faircoins",
-                        start_date=date,
-                        notes=notes,
-                        context_agent=from_agent,
-                    )
-                    exchange.save()
-                    transfer = Transfer(
-                        transfer_type=tt,
-                        exchange=exchange,
-                        transfer_date=date,
-                        name="Send Faircoins",
-                        notes=notes,
-                        context_agent=from_agent,
-                    )
-                    transfer.save()
-
-                # network_fee is subtracted from quantity
-                # so quantity is correct for the giving event
-                # but receiving event will get quantity - network_fee
-                state =  "new"
-                event = EconomicEvent(
-                    event_type = et_give,
-                    event_date = date,
-                    from_agent=from_agent,
-                    to_agent=to_agent,
-                    resource_type=resource.resource_type,
-                    resource=resource,
-                    digital_currency_tx_state = state,
-                    quantity = quantity,
-                    transfer=transfer,
-                    event_reference=address_end,
-                    description=notes,
-                    created_by=request.user,
-                    )
-                event.save()
-                if to_agent:
-                    quantity = quantity - Decimal(float(network_fee) / 1.e6)
-                    et_receive = EventType.objects.get(name="Receive")
-                    event = EconomicEvent(
-                        event_type = et_receive,
-                        event_date = date,
-                        from_agent=from_agent,
-                        to_agent=to_agent,
-                        resource_type=to_resource.resource_type,
-                        resource=to_resource,
-                        digital_currency_tx_state = state,
-                        quantity = quantity,
-                        transfer=transfer,
-                        event_reference=address_end,
-                        description=notes,
-                        created_by=request.user,
-                        )
-                    event.save()
-
-                return HttpResponseRedirect('/%s/%s/'
-                    % ('work/faircoin-history', resource.id))
-
-        return HttpResponseRedirect('/%s/%s/'
-                % ('work/manage-faircoin-account', resource.id))
-
-
-@login_required
-def faircoin_history(request, resource_id):
-    if not settings.USE_FAIRCOIN:
-        raise Http404
-    resource = get_object_or_404(EconomicResource, id=resource_id)
-    agent = get_agent(request)
-    wallet = faircoin_utils.is_connected()
-    confirmed_balance = None
-    unconfirmed_balance = None
-    if wallet:
-        if resource.is_wallet_address():
-            exchange_service = ExchangeService.get()
-            exchange_service.include_blockchain_tx_as_event(resource.owner(), resource)
-            try:
-                balances = faircoin_utils.get_address_balance(resource.digital_currency_address)
-                confirmed_balance = Decimal(balances[0]) / FAIRCOIN_DIVISOR
-                unconfirmed_balance =  Decimal(balances[0] + balances[1]) / FAIRCOIN_DIVISOR
-                unconfirmed_balance += resource.balance_in_tx_state_new()
-            except:
-                confirmed_balance = "Not accessible now"
-                unconfirmed_balance = "Not accessible now"
-        else:
-            wallet = False
-    event_list = resource.events.all()
-    init = {"quantity": resource.quantity,}
-    unit = resource.resource_type.unit
-
-    paginator = Paginator(event_list, 25)
-
-    page = request.GET.get('page')
-    try:
-        events = paginator.page(page)
-    except PageNotAnInteger:
-        # If page is not an integer, deliver first page.
-        events = paginator.page(1)
-    except EmptyPage:
-        # If page is out of range (e.g. 9999), deliver last page of results.
-        events = paginator.page(paginator.num_pages)
-    if resource.owner() == agent or resource.owner() in agent.managed_projects() or agent.is_staff():
-        return render(request, "work/faircoin_history.html", {
-            "resource": resource,
-            "agent": agent,
-            "confirmed_balance": confirmed_balance,
-            "unconfirmed_balance": unconfirmed_balance,
-            "unit": unit,
-            "events": events,
-            "wallet": wallet,
-        })
-    else:
-        return render(request, 'work/no_permission.html')
-
-
-
-@login_required
-def edit_faircoin_event_description(request, resource_id):
-    if not settings.USE_FAIRCOIN:
-        raise Http404
-    agent = get_agent(request)
-    resource = EconomicResource.objects.get(id=resource_id)
-    if request.method == "POST":
-        data = request.POST
-        evid = data['id']
-        ntext = data['value']
-        event = EconomicEvent.objects.get(id=evid)
-        event.description = ntext
-        event.save()
-        return HttpResponse("Ok", content_type="text/plain")
-    return HttpResponse("Fail", content_type="text/plain")
-
-
-
-
-
 
 #    M E M B E R S H I P
